add feedback field on feedback page + made the feedback menu on the home page less disruptive visually
